Toppings implemented.

+++ ProductManagementSystem/src/main/java/com/ferozkhandev/PizzaOrderingSystem/ProductManagementSystem/repositories/ToppingsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ToppingsRepository extends JpaRepository<Toppings, Integer> {
+public interface ToppingsRepository extends JpaRepository<Toppings, Long> {
 }
